requirements.txt
jinja2
fastapi
aiofiles
uvicorn
python-arango